<feat> 이미지 저장 수정, 필터 기능 구현 중

+++ src/main/java/com/project/compagnoserver/controller/ProductBoardController.java
@@ -46,16 +46,18 @@ public class ProductBoardController {
     private ProductBoardCommentService comment;
 
     @Value("${spring.servlet.multipart.location}")
-    private String uploadPath; // C:\\upload
+    private String uploadPath;
 
     // 게시물 등록
     @PostMapping("/productBoard")
     public ResponseEntity<ProductBoard> create(@RequestBody ProductBoardDTO dto) throws IOException {
         String saveName = null;
+        String saveFileName = null;
         // 메인 이미지 업로드
         if(dto.getProductMainFile() != null && !dto.getProductMainFile().isEmpty()) {
             String fileName = dto.getProductMainFile().getOriginalFilename();
             String uuid = UUID.randomUUID().toString();
+            saveFileName = "productBoard" + File.separator + uuid + "_" + fileName;
             saveName = uploadPath + File.separator + "productBoard" + File.separator + uuid + "_" + fileName;
             Path savePath = Paths.get(saveName);
             dto.getProductMainFile().transferTo(savePath);
@@ -64,7 +66,7 @@ public class ProductBoardController {
         // 게시판 작성
         ProductBoard vo = ProductBoard.builder()
                 .productBoardTitle(dto.getProductBoardTitle())
-                .productMainImage(saveName)
+                .productMainImage(saveFileName)
                 .productName(dto.getProductName())
                 .productPrice(dto.getProductPrice())
                 .productCategory(dto.getProductCategory())
@@ -83,6 +85,7 @@ public class ProductBoardController {
                 if (!fileName.isEmpty()) {
                     ProductBoardImage imgVo = new ProductBoardImage();
                     String uuid = UUID.randomUUID().toString();
+                    saveFileName = "productBoard" + File.separator + uuid + "_" + fileName;
                     saveName = uploadPath + File.separator + "productBoard" + File.separator + uuid + "_" + fileName;
                     Path savePath = Paths.get(saveName);
                     file.transferTo(savePath);
@@ -157,7 +160,7 @@ public class ProductBoardController {
 
     // 게시판 수정 (이미지 관련 수정 필요)
     @PutMapping("/productBoard")
-    public ResponseEntity<ProductBoard> update(ProductBoardDTO dto) throws IOException {
+    public ResponseEntity<ProductBoard> update(@RequestBody ProductBoardDTO dto) throws IOException {
 
 
         // 게시판 수정
@@ -176,7 +179,7 @@ public class ProductBoardController {
 
         // 기존 메인 이미지 삭제
         ProductBoard prev = productBoard.viewBoard(dto.getProductBoardCode());
-        if(prev.getProductMainImage() != null) {
+        if(prev != null &&prev.getProductMainImage() != null) {
             if (!prev.getProductMainImage().equals(dto.getMainImage()) || !dto.getProductMainFile().isEmpty()) {
                 File file = new File(prev.getProductMainImage());
                 file.delete();
@@ -184,16 +187,17 @@ public class ProductBoardController {
         }
 
         // 메인 이미지 업로드
-        if(!dto.getProductMainFile().isEmpty()) {
+        if(dto.getProductMainFile() != null && !dto.getProductMainFile().isEmpty()) {
             String fileName = dto.getProductMainFile().getOriginalFilename();
             String uuid = UUID.randomUUID().toString();
+            String saveFileName = "productBoard" + File.separator + uuid + "_" + fileName;
             String saveName = uploadPath + File.separator + "productBoard" + File.separator + uuid + "_" + fileName;
             Path savePath = Paths.get(saveName);
             dto.getProductMainFile().transferTo(savePath);
-            vo.setProductMainImage(saveName);
+            vo.setProductMainImage(saveFileName);
         }
 
-        if(prev.getProductMainImage() != null && dto.getProductMainFile().isEmpty()
+        if(prev != null && prev.getProductMainImage() != null && dto.getProductMainFile()!= null &&dto.getProductMainFile().isEmpty()
                 && prev.getProductMainImage().equals(dto.getMainImage())){
             String saveName = prev.getProductMainImage();
             vo.setProductMainImage(saveName);
@@ -220,13 +224,14 @@ public class ProductBoardController {
 
                 String fileName = file.getOriginalFilename();
                 String uuid = UUID.randomUUID().toString();
+                String saveFileName = "productBoard" + File.separator + uuid + "_" + fileName;
                 String saveName = uploadPath + File.separator + "productBoard" + File.separator + uuid + "_" + fileName;
 
                 Path savePath = Paths.get(saveName);
                 file.transferTo(savePath);
 
                 imgVo.setProductBoard(result);
-                imgVo.setProductImage(saveName);
+                imgVo.setProductImage(saveFileName);
 
                 productBoard.createImage(imgVo);
             }
@@ -239,7 +244,6 @@ public class ProductBoardController {
     // 게시판 추천, 추천 된 상태면 삭제
     @PostMapping("/productBoard/recommend")
     public ResponseEntity boardRecommend(@RequestBody ProductBoardRecommendDTO dto) {
-        log.info("dto : " + dto);
         ProductBoardRecommend vo = new ProductBoardRecommend();
         vo.setUser(User.builder().userId(dto.getUserId()).build());
         vo.setProductBoard(ProductBoard.builder().productBoardCode(dto.getProductBoardCode()).build());
@@ -254,8 +258,6 @@ public class ProductBoardController {
     public ResponseEntity boardBookmark(@RequestBody ProductBoardBookmarkDTO dto) {
         ProductBoardBookmark vo = new ProductBoardBookmark();
         vo.setUser(User.builder().userId(dto.getUserId()).build());
-        vo.setProductBoard(ProductBoard.builder().productBoardCode(dto.getProductBoardCode()).build());
-        log.info("vo : " + vo);
         productBoard.boardBookmark(vo);
 
         return ResponseEntity.ok().build();
@@ -276,7 +278,7 @@ public class ProductBoardController {
 
     // 게시판 검색, 조회
     @GetMapping("/public/productBoard/search")
-    public ResponseEntity<Page<ProductBoard>> searchBoard(ProductBoardSearchDTO dto, @RequestParam(name = "page", defaultValue = "1") int page) {
+    public ResponseEntity<Page<ProductBoard>> searchBoard(@ModelAttribute ProductBoardSearchDTO dto, @RequestParam(name = "page", defaultValue = "1") int page) {
         Pageable pageable = PageRequest.of(page - 1, 12);
         QProductBoard qProductBoard = QProductBoard.productBoard;
 
@@ -311,7 +313,6 @@ public class ProductBoardController {
                         .build())
                 .productCommentCode(dto.getProductCommentCode())
                 .productCommentContent(dto.getProductCommentContent())
-                .productParentCode(dto.getProductParentCode())
                 .user(User.builder()
                         .userId(userInfo().getUserId())
                         .build())
